% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{coef.peer_cra}
\alias{coef.peer_cra}
\title{Extract coefficients from a fitted peer-GMM object.}
\usage{
\method{coef}{peer_cra}(object, ...)
}
\arguments{
\item{object}{a \code{peer_cra} object.}

\item{...}{ignored.}
}
\description{
Extract coefficients from a fitted peer-GMM object.
}
