% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{fitted.peer_cra}
\alias{fitted.peer_cra}
\title{Fitted values from a fitted peer-GMM object.}
\usage{
\method{fitted}{peer_cra}(object, ...)
}
\arguments{
\item{object}{a \code{peer_cra} object.}

\item{...}{ignored.}
}
\value{
numeric vector \eqn{\hat Y_i = Y_i - r_i}.
}
\description{
Fitted values from a fitted peer-GMM object.
}
