% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{nobs.peer_cra}
\alias{nobs.peer_cra}
\title{Number of observations in a fitted peer-GMM object.}
\usage{
\method{nobs}{peer_cra}(object, ...)
}
\arguments{
\item{object}{a \code{peer_cra} object.}

\item{...}{ignored.}
}
\description{
Number of observations in a fitted peer-GMM object.
}
