% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{residuals.peer_cra}
\alias{residuals.peer_cra}
\title{Residuals from a fitted peer-GMM object.}
\usage{
\method{residuals}{peer_cra}(object, ...)
}
\arguments{
\item{object}{a \code{peer_cra} object.}

\item{...}{ignored.}
}
\value{
numeric vector of length n.
}
\description{
Structural-equation residuals: \eqn{r_{i} = Y_i - \hat\lambda (W Y)_i -
Z_i' \hat\beta - \hat\alpha_{r(i)}}, where \eqn{\hat\alpha_r} is the urn
mean of \eqn{Y - \hat\lambda W Y - Z \hat\beta}.
}
