% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{vcov.peer_cra}
\alias{vcov.peer_cra}
\title{Extract the asymptotic variance-covariance matrix.}
\usage{
\method{vcov}{peer_cra}(object, ...)
}
\arguments{
\item{object}{a \code{peer_cra} object.}

\item{...}{ignored.}
}
\description{
Extract the asymptotic variance-covariance matrix.
}
